Update User.find_by_email to SQLA
